requirements.txt
opencv-python
torch>=2.2.0
torchvision>=0.17.0
streamlit
pillow
numpy
easyocr
